compiler: Move SyncOp hierarchy from types/ to ir/
