+++ devito/ir/support/__init__.py
@@ -2,5 +2,6 @@ from .vector import *  # noqa
 from .basic import *  # noqa
 from .space import *  # noqa
 from .guards import *  # noqa
+from .syncs import *  # noqa
 from .utils import *  # noqa
 from .properties import *  # noqa
