+++ devito/types/parallel.py
@@ -7,25 +7,22 @@ identifying a device attached to a node).
 """
 
 import os
-from collections import defaultdict
 from ctypes import c_void_p
 
 from cached_property import cached_property
 import numpy as np
-import sympy
 
 from devito.exceptions import InvalidArgument
 from devito.parameters import configuration
-from devito.tools import Pickable, as_list, as_tuple, dtype_to_cstr, filter_ordered
+from devito.tools import as_list, as_tuple, dtype_to_cstr
 from devito.types.array import Array, ArrayObject
 from devito.types.basic import Scalar, Symbol
 from devito.types.dimension import CustomDimension
 from devito.types.misc import Pointer, VolatileInt, c_volatile_int_p
 
 __all__ = ['NThreads', 'NThreadsNested', 'NThreadsNonaffine', 'NThreadsBase',
-           'DeviceID', 'ThreadID', 'Lock', 'WaitLock', 'WithLock', 'FetchUpdate',
-           'FetchPrefetch', 'PrefetchUpdate', 'WaitPrefetch', 'Delete', 'PThreadArray',
-           'SharedData', 'NPThreads', 'DeviceRM', 'DevicePointer', 'normalize_syncs']
+           'DeviceID', 'ThreadID', 'Lock', 'PThreadArray', 'SharedData',
+           'NPThreads', 'DeviceRM', 'DevicePointer']
 
 
 class NThreadsBase(Scalar):
@@ -239,149 +236,6 @@ class Lock(Array):
 
     # Pickling support
     _pickle_kwargs = Array._pickle_kwargs + ['target']
-
-
-class SyncOp(sympy.Expr, Pickable):
-
-    is_SyncLock = False
-    is_SyncData = False
-
-    is_WaitLock = False
-    is_WithLock = False
-    is_Fetch = False
-
-    def __new__(cls, handle):
-        obj = sympy.Expr.__new__(cls, handle)
-        obj.handle = handle
-        return obj
-
-    def __str__(self):
-        return "%s<%s>" % (self.__class__.__name__, self.handle)
-
-    __repr__ = __str__
-
-    def __eq__(self, other):
-        return type(self) == type(other) and self.args == other.args
-
-    def __hash__(self):
-        return hash((type(self), self.args))
-
-    # Pickling support
-    _pickle_args = ['handle']
-    __reduce_ex__ = Pickable.__reduce_ex__
-
-
-class SyncLock(SyncOp):
-
-    is_SyncLock = True
-
-    @property
-    def lock(self):
-        return self.handle.function
-
-    @property
-    def target(self):
-        return self.lock.target
-
-
-class SyncData(SyncOp):
-
-    is_SyncData = True
-
-    def __new__(cls, dim, size, function, fetch, ifetch, fcond,
-                pfetch=None, pcond=None, target=None, tstore=None):
-        obj = sympy.Expr.__new__(cls, dim, size, function, fetch, ifetch, fcond,
-                                 pfetch, pcond, target, tstore)
-
-        # fetch -> the input Function fetch index, e.g. `time`
-        # ifetch -> the input Function initialization index, e.g. `time_m`
-        # pfetch -> the input Function prefetch index, e.g. `time+1`
-        # tstore -> the target Function store index, e.g. `sb1`
-
-        # fcond -> the input Function fetch condition, e.g. `time_m <= time_M`
-        # pcond -> the input Function prefetch condition, e.g. `time + 1 <= time_M`
-
-        obj.dim = dim
-        obj.size = size
-        obj.function = function
-        obj.fetch = fetch
-        obj.ifetch = ifetch
-        obj.fcond = fcond
-        obj.pfetch = pfetch
-        obj.pcond = pcond
-        obj.target = target
-        obj.tstore = tstore
-
-        return obj
-
-    def __str__(self):
-        return "%s<%s->%s:%s:%d>" % (self.__class__.__name__, self.function,
-                                     self.target, self.dim, self.size)
-
-    __repr__ = __str__
-
-    __hash__ = sympy.Basic.__hash__
-
-    @property
-    def dimensions(self):
-        return self.function.dimensions
-
-    # Pickling support
-    _pickle_args = ['dim', 'size', 'function', 'fetch', 'ifetch', 'fcond']
-    _pickle_kwargs = ['pfetch', 'pcond', 'target', 'tstore']
-    __reduce_ex__ = Pickable.__reduce_ex__
-
-
-class WaitLock(SyncLock):
-    is_WaitLock = True
-
-
-class WithLock(SyncLock):
-    is_WithLock = True
-
-
-class FetchPrefetch(SyncData):
-    is_Fetch = True
-
-
-class FetchUpdate(SyncData):
-    is_Fetch = True
-
-
-class PrefetchUpdate(SyncData):
-    is_Fetch = True
-
-
-class WaitPrefetch(SyncData):
-    pass
-
-
-class Delete(SyncData):
-    pass
-
-
-def normalize_syncs(*args):
-    if not args:
-        return
-    if len(args) == 1:
-        return args[0]
-
-    syncs = defaultdict(list)
-    for _dict in args:
-        for k, v in _dict.items():
-            syncs[k].extend(v)
-
-    syncs = {k: filter_ordered(v) for k, v in syncs.items()}
-
-    for v in syncs.values():
-        waitlocks = [i for i in v if i.is_WaitLock]
-        withlocks = [i for i in v if i.is_WithLock]
-
-        if waitlocks and withlocks:
-            # We do not allow mixing up WaitLock and WithLock ops
-            raise ValueError("Incompatible SyncOps")
-
-    return syncs
 
 
 class DeviceSymbol(Scalar):
